fix: seperated logic of styling in Panel
